requirements.txt
aiohttp==3.6.2
appdirs==1.4.4
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
click>=8.0.3
gunicorn==20.0.4
idna==2.9
mqtt==0.0.1
multidict==4.7.5
paho-mqtt==1.5.0
pathspec>=0.9.0
psycopg2-binary>=2.9.3
regex==2020.5.7
toml==0.10.0
typed-ast==1.4.1
yarl==1.4.2
pyjwt
rsa

flake8>=4.0.1
isort>=5.10.1
black>=22.1.0
pylint>=2.12.2
mypy>=0.931
